+++ app/Exceptions/Handler.php
@@ -40,8 +40,11 @@ class Handler extends ExceptionHandler
     public function render($request, \Exception $e)
     {
         if ($e instanceof RedirectException) {
+            return redirect($e->getTarget())->with('messages', $e->toArray());
+        }
 
-            return redirect($e->getTarget())->with('error', $e->getError());
+        if (!$this->isHttpException($e) && !config('app.debug')) {
+            return response()->view('errors.500', [], 500);
         }
 
         return parent::render($request, $e);