+++ resources/views/errors/403.blade.php
@@ -1,0 +1,5 @@
+@extends('master')
+
+@section('content')
+    <h1>You don't have permission to access this area.</h1>
+@endsection