+++ resources/views/includes/page/version.blade.php
@@ -1,0 +1,2 @@
+{{--GENERATED: 2015-08-25 11:58:28--}}
+version: beta 0.4.0-7-g5d05 (download-feature)