+++ CHANGELOG.md
@@ -1,5 +1,71 @@
 ## Features
+- Release version in footer
+
+## Bug Fixes
+- Email notifications for converted application
+
+# 0.8.0
+2015-08-21
+
+## Features
+- Limit actions on *Applications List*
+- Middleware test
+
+##Bug Fixes
+- Removed non working links
+- Fixed two middleware
+- Postcode filter on *Applications List*
+- Currency filters on *Applications List*
+
+# 0.7.0
+2015-08-19
+
+## Features
+- Removed footer contents
+- Redeveloped Application Process
+- Change layout to fluid
+- Error pages for 404 and 50X
+
+## Bug Fixes
+- Login page errors
+- Minor fixes in view
+
+# 0.6.1
+2015-08-18
+
+# 0.6.0
+2015-08-18
+
+## Features
+- Actions on view pages have been changed to be included inside views
+- Models linked to each other
+- Active flag has been changed to a slider on edit page
+- Confirmation screens have been made 'pretty'
+- Messages moved to master
+- Logo has been changed
+- Validity period has been added to installation. Can be changed on edit page.
+- Validity period of application is dependent on the validity period in installations
+- Requester and location added to application
+- Return page
+- Menu structure changed
+- Menu items now have a scope depending on a users permissions
+- Middleware for available installation
+- Added columns & filters for applications list.
+
+## Bug Fixes
+- 'Gravatar' bug fixed
+- `SettlementGateway` fixes
+
+# 0.5.0
+2015-08-14
+
+## Features
 - Changed layout of dashboard
+- Set `Permissions`. Attach `Permissions` to `Routes`
+- `AuthoriseRole` middleware
+- Protect *Super Administrator* role
+- Protect against `user` self delete
+- Test improvements
 
 # 0.4.0
 2015-08-11